services/worker/requirements.txt
nats-py==2.7.2
APScheduler==3.10.4
requests==2.32.3
